requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
django-fastdev==1.13.0
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
iommi==7.20.0
Markdown==3.10
pillow==12.0.0
pyparsing==3.2.5
sqlparse==0.5.3
whitenoise==6.11.0
